requirements.txt
Django
Unipath
South
psycopg2
markdown
djangorestframework
django-oauth-toolkit
djangorestframework-jwt
python-social-auth
django-cors-headers
